+++ spec/logging_spec.rb
@@ -46,155 +46,200 @@ describe 'openstack-common::logging' do
             'ignore.key.1' => 'ignore.value.1',
             'ignore.key.2' => 'ignore.value.2'
           }
-          [
-            /^\[loggers\]$/,
-            /^keys=.*ignore_key_1,ignore_key_2$/
-          ].each do |content|
-            expect(chef_run).to render_file(file.name).with_content(content)
-          end
+          expect(chef_run).to render_config_file(file.name)
+              .with_section_content('loggers', /^keys=.*ignore_key_1,ignore_key_2$/)
         end
 
         it 'adds specific logger ignore block' do
           node.set['openstack']['logging']['ignore'] = {
             'test.nova.api.openstack.wsgi' => 'WARNING'
           }
-          content = [
-            '[logger_test_nova_api_openstack_wsgi]',
+          [
             'level = WARNING',
             'handlers = prod,debug',
             'qualname = test.nova.api.openstack.wsgi'
-          ]
-          expect(chef_run).to render_file(file.name).with_content(build_section(content))
+          ].each do |line|
+            expect(chef_run).to render_config_file(file.name)
+                .with_section_content('logger_test_nova_api_openstack_wsgi', line)
+          end
         end
       end
 
       context 'loggers' do
         it 'adds default loggers' do
-          [
-            ['[loggers]',
-             'keys=root,ceilometer,cinder,glance,horizon,keystone,nova,'\
-               'neutron,swift,trove,amqplib,sqlalchemy,boto,suds,eventletwsgi,'\
-               'nova_api_openstack_wsgi,nova_osapi_compute_wsgi_server'],
-            ['[logger_root]',
-             'level=NOTSET',
-             'handlers=devel'],
-            ['[logger_ceilometer]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=ceilometer'],
-            ['[logger_cinder]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=cinder'],
-            ['[logger_glance]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=glance'],
-            ['[logger_horizon]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=horizon'],
-            ['[logger_keystone]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=keystone'],
-            ['[logger_nova]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=nova'],
-            ['[logger_neutron]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=neutron'],
-            ['[logger_swift]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=swift'],
-            ['[logger_trove]',
-             'level=DEBUG',
-             'handlers=prod,debug',
-             'qualname=trove'],
-            ['[logger_amqplib]',
-             'level=WARNING',
-             'handlers=stderr',
-             'qualname=amqplib'],
-            ['[logger_sqlalchemy]',
-             'level=WARNING',
-             'handlers=stderr',
-             'qualname=sqlalchemy'],
-            ['[logger_boto]',
-             'level=WARNING',
-             'handlers=stderr',
-             'qualname=boto'],
-            ['[logger_suds]',
-             'level=INFO',
-             'handlers=stderr',
-             'qualname=suds'],
-            ['[logger_eventletwsgi]',
-             'level=WARNING',
-             'handlers=stderr',
-             'qualname=eventlet.wsgi.server'],
-            ['[logger_nova_api_openstack_wsgi]',
-             'level=WARNING',
-             'handlers=prod,debug',
-             'qualname=nova.api.openstack.wsgi'],
-            ['[logger_nova_osapi_compute_wsgi_server]',
-             'level=WARNING',
-             'handlers=prod,debug',
-             'qualname=nova.osapi_compute.wsgi.server']
-          ].each do |content|
-            expect(chef_run).to render_file(file.name).with_content(build_section(content))
+          {
+            'loggers' =>
+              [
+                'keys=root,ceilometer,cinder,glance,horizon,keystone,nova,'\
+                'neutron,swift,trove,amqplib,sqlalchemy,boto,suds,eventletwsgi,'\
+                'nova_api_openstack_wsgi,nova_osapi_compute_wsgi_server'
+              ],
+            'logger_root' =>
+              [
+                'level=NOTSET',
+                'handlers=devel'
+              ],
+            'logger_ceilometer' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=ceilometer'
+              ],
+            'logger_cinder' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=cinder'
+              ],
+            'logger_glance' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=glance'
+              ],
+            'logger_horizon' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=horizon'
+              ],
+            'logger_keystone' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=keystone'
+              ],
+            'logger_nova' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=nova'
+              ],
+            'logger_neutron' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=neutron'
+              ],
+            'logger_swift' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=swift'
+              ],
+            'logger_trove' =>
+              [
+                'level=DEBUG',
+                'handlers=prod,debug',
+                'qualname=trove'
+              ],
+            'logger_amqplib' =>
+              [
+                'level=WARNING',
+                'handlers=stderr',
+                'qualname=amqplib'
+              ],
+            'logger_sqlalchemy' =>
+              [
+                'level=WARNING',
+                'handlers=stderr',
+                'qualname=sqlalchemy'
+              ],
+            'logger_boto' =>
+              [
+                'level=WARNING',
+                'handlers=stderr',
+                'qualname=boto'
+              ],
+            'logger_suds' =>
+              [
+                'level=INFO',
+                'handlers=stderr',
+                'qualname=suds'
+              ],
+            'logger_eventletwsgi' =>
+              [
+                'level=WARNING',
+                'handlers=stderr',
+                'qualname=eventlet.wsgi.server'
+              ],
+            'logger_nova_api_openstack_wsgi' =>
+              [
+                'level=WARNING',
+                'handlers=prod,debug',
+                'qualname=nova.api.openstack.wsgi'
+              ],
+            'logger_nova_osapi_compute_wsgi_server' =>
+              [
+                'level=WARNING',
+                'handlers=prod,debug',
+                'qualname=nova.osapi_compute.wsgi.server'
+              ]
+          }.each do |section, content|
+            content.each do |line|
+              expect(chef_run).to render_config_file(file.name).with_section_content(section, line)
+            end
           end
         end
       end
 
       context 'formatters' do
         it 'adds default formatters' do
-          [
-            ['[formatters]',
-             'keys=normal,normal_with_name,debug,syslog_with_name,syslog_debug'],
-            ['[formatter_normal]',
-             'format=%(asctime)s %(levelname)s %(message)s'],
-            ['[formatter_normal_with_name]',
-             'format=[%(name)s]: %(asctime)s %(levelname)s %(message)s'],
-            ['[formatter_debug]',
-             'format=[%(name)s]: %(asctime)s %(levelname)s %(module)s.%(funcName)s %(message)s'],
-            ['[formatter_syslog_with_name]',
-             'format=%(name)s: %(levelname)s %(message)s'],
-            ['[formatter_syslog_debug]',
-             'format=%(name)s: %(levelname)s %(module)s.%(funcName)s %(message)s']
-          ].each do |content|
-            expect(chef_run).to render_file(file.name).with_content(build_section(content))
+          {
+            'formatters' =>
+              'keys=normal,normal_with_name,debug,syslog_with_name,syslog_debug',
+            'formatter_normal' =>
+              'format=%(asctime)s %(levelname)s %(message)s',
+            'formatter_normal_with_name' =>
+              'format=[%(name)s]: %(asctime)s %(levelname)s %(message)s',
+            'formatter_debug' =>
+              'format=[%(name)s]: %(asctime)s %(levelname)s %(module)s.%(funcName)s %(message)s',
+            'formatter_syslog_with_name' =>
+              'format=%(name)s: %(levelname)s %(message)s',
+            'formatter_syslog_debug' =>
+              'format=%(name)s: %(levelname)s %(module)s.%(funcName)s %(message)s'
+          }.each do |section, content|
+            expect(chef_run).to render_config_file(file.name).with_section_content(section, content)
           end
         end
       end
 
       context 'handlers' do
         it 'adds default handlers' do
-          [
-            ['[handlers]',
-             'keys=stderr,devel,prod,debug'],
-            ['[handler_stderr]',
-             'args=(sys.stderr,)',
-             'class=StreamHandler',
-             'formatter=debug'],
-            ['[handler_devel]',
-             'args=(sys.stdout,)',
-             'class=StreamHandler',
-             'formatter=debug',
-             'level=NOTSET'],
-            ['[handler_prod]',
-             "args=(('/dev/log'), handlers.SysLogHandler.LOG_LOCAL0)",
-             'class=handlers.SysLogHandler',
-             'formatter=syslog_with_name',
-             'level=INFO'],
-            ['[handler_debug]',
-             "args=(('/dev/log'), handlers.SysLogHandler.LOG_LOCAL1)",
-             'class=handlers.SysLogHandler',
-             'formatter=syslog_debug',
-             'level=DEBUG']
-          ].each do |content|
-            expect(chef_run).to render_file(file.name).with_content(build_section(content))
+          {
+            'handlers' =>
+              ['keys=stderr,devel,prod,debug'],
+            'handler_stderr' =>
+              [
+                'args=(sys.stderr,)',
+                'class=StreamHandler',
+                'formatter=debug'
+              ],
+            'handler_devel' =>
+              [
+                'args=(sys.stdout,)',
+                'class=StreamHandler',
+                'formatter=debug',
+                'level=NOTSET'
+              ],
+            'handler_prod' =>
+              [
+                "args=(('/dev/log'), handlers.SysLogHandler.LOG_LOCAL0)",
+                'class=handlers.SysLogHandler',
+                'formatter=syslog_with_name',
+                'level=INFO'
+              ],
+            'handler_debug' =>
+              [
+                "args=(('/dev/log'), handlers.SysLogHandler.LOG_LOCAL1)",
+                'class=handlers.SysLogHandler',
+                'formatter=syslog_debug',
+                'level=DEBUG'
+              ]
+          }.each do |section, content|
+            content.each do |line|
+              expect(chef_run).to render_config_file(file.name).with_section_content(section, line)
+            end
           end
         end
       end