+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of cookbook-openstack-com
 * Add OS_VOLUME_API_VERSION for openrc
 * Removed the hardcoded PPA reference to "Precise"
 * Added an optional automatic apt-get update for Debian based repos
+* Add a new custom matcher render_config_file to test ini file content
 
 ## 10.1.0
 * Adding identity admin bind host endpoint to allow flexibility and consistency